Completes Backend

+++ backend/controllers/product.controller.js
@@ -1,4 +1,4 @@
-import Product from '../models/product.model';
+import Product from '../models/product.model.js';
 import mongoose from 'mongoose';
 
 export const getProducts = async (req, res) => {
@@ -31,12 +31,18 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: 'Invalid Product ID' });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: 'Product deleted' });
   } catch (error) {
     console.error(`Error deleting Product: ${error.message}`);
-    res.status(404).json({ success: false, message: 'Product not found' });
+    res.status(500).json({ success: false, message: 'Server Error' });
   }
 };
